requirements.txt
torch>=2.0.0
torchvision>=0.15.0
xgboost>=2.0.0
flwr[simulation]>=1.5.0
numpy>=1.21.0
pandas>=1.0.0
Pillow>=9.0.0
scikit-learn>=1.0.0
captum>=0.6.0

